+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [1.2.7] / 2023-04-15
+### Features
+- TestCase with string apear with `"` in the name.
+- TestCase with null apear with `null` in the name.
 ### Fixed
 - Fix #21 issue with `Nullable` check.
 ### Tests